% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adjMatrix}
\alias{adjMatrix}
\title{Draw adjacency matrix between colonies}
\usage{
adjMatrix(locations, maxDistance)
}
\arguments{
\item{locations}{x, y numeric locations of each colony.}

\item{maxDistance}{For kernels with an upper threshold}
}
\description{
For unweighted networks, draw an adjacency matrix showing which colonies are connected and which aren't
}
\seealso{
Other Networks: \code{\link{checkIfGiant}};
  \code{\link{edgeList}}; \code{\link{netDegree}};
  \code{\link{netStrength}}; \code{\link{numComponents}};
  \code{\link{popMatrix}}; \code{\link{weightMatrix}}
}

