% Generated by roxygen2 (4.0.2): do not edit by hand
\name{circleColony}
\alias{circleColony}
\title{Place colonies in a circle. Most good for viz when all colonies are connected etc..}
\usage{
circleColony(N, space)
}
\arguments{
\item{N}{Number of colonies. A single, positive integer.}

\item{space}{The size of space as given by the length of one side of square space.}
}
\description{
Radial placement of colonies in two dimensions.
}
\seealso{
Other colony.placement: \code{\link{uniformColony}}
}

