% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coinfectionR}
\alias{coinfectionR}
\title{Calculate coinfection rates for each colony.}
\usage{
coinfectionR(pop, t)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}

\item{t}{Time. An integer giving the event number (i.e. the 1000th event).}
}
\value{
nColonies * 2*n - n - 1 (i.e. no of coinfected classed)
}
\description{
Calculate coinfection rates for each colony.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{deathR}}; \code{\link{dispersalR}};
  \code{\link{infectionR}}; \code{\link{infectionTrans}};
  \code{\link{transRates}}
}

