% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{dispersalR}
\alias{dispersalR}
\title{Calculate dispersal rates for each colony.}
\usage{
dispersalR(pop, t)
}
\arguments{
\item{pop}{A population object}

\item{t}{Time step. Note the population should be AT time t, going to t+1.}
}
\description{
Calculate starting dispersal rate for each colony. Later functions recalculate relevant rates after an event.
}

