% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{edgeList}
\alias{edgeList}
\title{Create a list of edges with a column of weights}
\usage{
edgeList(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Create a list of edges with a column of weights
}

