% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeDistributionModels.R
\name{exponentialPop}
\alias{exponentialPop}
\title{Create exponential colony size distribution.}
\usage{
exponentialPop(nColonies, meanColonySize)
}
\arguments{
\item{nColonies}{The number of colonies. A single integer.}

\item{meanColonySize}{Mean colony size.}
}
\description{
Create random colony size distirbution from exponential distribution.
}

