% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findDisDistr}
\alias{findDisDistr}
\title{Calculate mean disease distributions (how many individuals infected with each disease) at the end of a simulation}
\usage{
findDisDistr(pop, final = 10)
}
\arguments{
\item{pop}{A population object from makePop()}

\item{final}{Integer giving the number of recorded population states to average over. Note that this is the number of recorded events, not actual events so will cover a period of final * sample.}
}
\description{
Calculate mean disease distributions (how many individuals infected with each disease) at the end of a simulation
}

