% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{findDiseaseAdded}
\alias{findDiseaseAdded}
\title{Make vector of which disease is added for each coinfection transitions in pop$transitions}
\usage{
findDiseaseAdded(pop)
}
\arguments{
\item{pop}{A population object}
}
\description{
Make vector of which disease is added for each coinfection transitions in pop$transitions
}

