% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{infectionTrans}
\alias{infectionTrans}
\title{Find possible possible infection transitions.}
\usage{
infectionTrans(pop)
}
\arguments{
\item{pop}{A population object}
}
\description{
Find possible possible infection transitions.
}

