% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePop.R
\name{initTransitions}
\alias{initTransitions}
\title{Build the transition data.frame which will have rates populated later.}
\usage{
initTransitions(pop)
}
\arguments{
\item{pop}{A population object}
}
\description{
Build the transition data.frame which will have rates populated later.
}

