% Generated by roxygen2 (4.0.2): do not edit by hand
\name{numComponents}
\alias{numComponents}
\title{Count the number of components in population network}
\usage{
numComponents(W)
}
\arguments{
\item{W}{A weighted adjacency matrix}
}
\description{
If sections of the network are completely unconnected to the rest of the population this has odd effects for an analysis.
  Therefore check how many components there are.
}
\seealso{
\code{\link{checkIfGiant}}

Other Networks: \code{\link{adjMatrix}};
  \code{\link{checkIfGiant}}; \code{\link{edgeList}};
  \code{\link{netDegree}}; \code{\link{netStrength}};
  \code{\link{popMatrix}}; \code{\link{weightMatrix}}
}

