% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pClass}
\alias{pClass}
\title{Plot the total number of individuals in each disease class}
\usage{
pClass(pop, start = 1, end = NULL, S = FALSE, nPath = TRUE, o = FALSE)
}
\arguments{
\item{pop}{A population object}

\item{o}{Logical whether to plot from the origin or not}
}
\description{
Plot the total number of individuals in each disease class
}

