% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pCol}
\alias{pCol}
\title{Plot the number of individuals by number of infections and colony}
\usage{
pCol(pop, start = 1, end = NULL, S = TRUE, nPath = TRUE, o = FALSE)
}
\arguments{
\item{pop}{A MetapopEpi class object}

\item{start}{What event to start plotting from.}

\item{end}{What event to end plotting from. If NULL, plot until end.}

\item{S}{Logical, controls whether or not to include the susceptible population.}

\item{nPath}{Logical, Colour the plot by number of pathogens or by class number}

\item{o}{Logical. Truncate y-axis.}
}
\description{
Plot the number of individuals by number of infections and colony
}

