% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pDis}
\alias{pDis}
\alias{pDisold}
\title{Plot the total number of infected individuals with each disease}
\usage{
pDisold(pop, start = 1, end = NULL, o = FALSE)

pDis(pop, start = 1, end = NULL, o = FALSE)
}
\arguments{
\item{pop}{A population object}

\item{start}{What event to start plotting from.}

\item{end}{What event to end plotting from. If NULL, plot until end.}

\item{o}{Logical whether to plot from the origin or not}
}
\description{
Plot the total number of infected individuals with each disease

Plot the total number of infected individuals with each disease
}

