% Generated by roxygen2 (4.0.1): do not edit by hand
\name{popHeat}
\alias{popHeat}
\title{A nice heat map for weighted adjacency matrices}
\usage{
popHeat(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Plot showing the weight of edges by colouring the adjacency matrix.
}
\seealso{
Other viz: \code{\link{colonyHeat}},
  \code{\link{plotColonyNet}}, \code{\link{plotColonyNet}}
}

