% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_MetapopEpi.R
\name{print.MetapopEpi}
\alias{print.MetapopEpi}
\title{Print method for objects of class MetapopEpi}
\usage{
\method{print}{MetapopEpi}(p)
}
\description{
Print method for objects of class MetapopEpi
}

