% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pSIR}
\alias{pSIR}
\title{Plot the total number of individuals in each disease class}
\usage{
pSIR(pop, start = 1, end = NULL, S = TRUE, nPath = TRUE, o = FALSE)
}
\arguments{
\item{pop}{A MetapopEpi class object}

\item{start}{What event to start plotting from.}

\item{end}{What event to end plotting from. If NULL, plot until end.}

\item{S}{Logical, controls whether or not to include the susceptible population.}

\item{nPath}{Logical, Colour the plot by number of pathogens or by class number}

\item{o}{Logical. Truncate y-axis.}
}
\description{
Plot the total number of individuals in each disease class
}

