% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{randEvent}
\alias{randEvent}
\title{Perform a random event}
\usage{
randEvent(pop, t, tMod)
}
\arguments{
\item{pop}{A population object}

\item{t}{Time step. Note the population should be AT time t, going to t+1.}

\item{tMod}{Time step modulus the \code{sample} argument in \code{makePop}.}
}
\description{
Randomly selects an event weighted by their rates, runs the event and then calculates a waiting time.
}

