% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seedPathogen}
\alias{seedPathogen}
\title{Add single individuals infected with a number of pathogens}
\usage{
seedPathogen(pop, pathogens = 1, n = 1)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}

\item{pathogens}{Character vector of which pathogens should be seeded}
}
\value{
An update object of same structure as from makePop
}
\description{
Allows addition of an infected individual of 1 or more pathogens.
}
\examples{
p <- makePop()
u <- seedPathogen(p, pathogens = 1:p$parameters['nPathogens'])
}

