% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transRates}
\alias{transRates}
\title{Calculate new transition rates.}
\usage{
transRates(pop, t)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}

\item{t}{Time. An integer giving the event number (i.e. the 1000th event).}
}
\description{
Calculate new transition rates.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{deathR}};
  \code{\link{dispersalR}}; \code{\link{findDiseaseAdded}};
  \code{\link{infectionR}}; \code{\link{infectionTrans}}
}

