% Generated by roxygen2 (4.0.1): do not edit by hand
\name{uniformColony}
\alias{uniformColony}
\title{Make a matrix of N colony locations placed by a spatial poisson process.}
\usage{
uniformColony(N, space)
}
\arguments{
\item{N}{Number of colonies. A single, positive integer.}

\item{space}{The size of space as given by the length of one side of square space.}
}
\description{
Uniformally random placement of colonies in two dimensions.
}
\seealso{
Other colony.placement: \code{\link{circleColony}}
}

