% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{unweightedKern}
\alias{unweightedKern}
\title{Unweighted kernel for calculating a weight matrix from a distance matrix}
\usage{
unweightedKern(distMatrix, thresh)
}
\arguments{
\item{distMatrix}{A distance matrix of the euclidean distances between each pair of colonies}

\item{thresh}{A maximum distance above which colonies are not connected.}
}
\description{
Creates an unweighted network. As long as distance < theshold, the weight is 1.
}

