% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weightMatrix}
\alias{weightMatrix}
\title{Draw weighted adjacency matrix between colonies}
\usage{
weightMatrix(locations, maxDistance, kern)
}
\arguments{
\item{locations}{x, y numeric locations of each colony.}

\item{maxDistance}{For kernels with an upper threshold}

\item{kern}{Character, giving the name of the distance kernel to be used.}
}
\description{
For weighted networks, draw an adjacency matrix showing which colonies are connected and how strongly connected they are
  depending on a kernel that defines how connection strength decreases with distance.
}
\details{
Many kernels have a maximum threshold above which colonies aren't connected. This simplified things as otherwise all colonies
  would be connected to some extent. This threshold is defined with the argument maxDistance.
}
\seealso{
Other Networks: \code{\link{adjMatrix}};
  \code{\link{checkIfGiant}}; \code{\link{edgeList}};
  \code{\link{netDegree}}; \code{\link{netStrength}};
  \code{\link{numComponents}}; \code{\link{popMatrix}}
}

