% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{adjMatrix}
\alias{adjMatrix}
\title{Draw adjacency matrix between colonies}
\usage{
adjMatrix(locations, maxDistance)
}
\arguments{
\item{locations}{x, y numeric locations of each colony.}

\item{maxDistance}{For kernels with an upper threshold}
}
\description{
For unweighted networks, draw an adjacency matrix showing which colonies are connected and which aren't
}

