% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkIfGiant}
\alias{checkIfGiant}
\title{Is the network one giant component or multiple small separate networks.}
\usage{
checkIfGiant(A)
}
\arguments{
\item{A}{An adjacency matrix}
}
\description{
If sections of the network are completely unconnected to the rest of the population this has odd effects for an analysis.
  Therefore check how many components there are.
}
\seealso{
\code{\link{numComponents}}

Other Networks: \code{\link{adjMatrix}};
  \code{\link{edgeList}}; \code{\link{netDegree}};
  \code{\link{netStrength}}; \code{\link{numComponents}};
  \code{\link{popMatrix}}; \code{\link{weightMatrix}}
}

