% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coinfectionR}
\alias{coinfectionR}
\title{Calculate coinfection rates for each colony.}
\usage{
coinfectionR(pop, t)
}
\arguments{
\item{pop}{A population object}

\item{t}{Time step. Note the population should be AT time t, going to t+1.}
}
\value{
nColonies * something. Grouped by to class, then from class
}
\description{
Calculate coinfection rates for each colony.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{deathR}}; \code{\link{dispersalR}};
  \code{\link{findDiseaseAdded}}; \code{\link{infectionR}},
  \code{\link{infectionR}}, \code{\link{recoveryR}};
  \code{\link{infectionTrans}};
  \code{\link{initTransitions}}; \code{\link{transRates}}
}

