% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{coinfectionR}
\alias{coinfectionR}
\title{Calculate coinfection rates for each colony.}
\usage{
coinfectionR(pop, t)
}
\arguments{
\item{pop}{A population object}

\item{t}{Time step. Note the population should be AT time t, going to t+1.}
}
\value{
nColonies * something. Grouped by to class, then from class
}
\description{
Calculate coinfection rates for each colony.
}

