% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deathR}
\alias{deathR}
\title{Calculate death rates for each colony.}
\usage{
deathR(pop, t)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}

\item{t}{Time. An integer giving the event number (i.e. the 1000th event).}
}
\description{
Calculate starting death rate for each colony. Later functions recalculate relevant rates after an event.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{dispersalR}};
  \code{\link{findDiseaseAdded}}; \code{\link{infectionR}};
  \code{\link{infectionTrans}};
  \code{\link{initTransitions}}; \code{\link{transRates}}
}

