% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dispersalR}
\alias{dispersalR}
\title{Calculate dispersal rates for each colony.}
\usage{
dispersalR(pop, t)
}
\arguments{
\item{pop}{A population object}

\item{t}{Time step. Note the population should be AT time t, going to t+1.}
}
\description{
Calculate starting dispersal rate for each colony. Later functions recalculate relevant rates after an event.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{deathR}};
  \code{\link{findDiseaseAdded}}; \code{\link{infectionR}},
  \code{\link{infectionR}}, \code{\link{recoveryR}};
  \code{\link{infectionTrans}};
  \code{\link{initTransitions}}; \code{\link{transRates}}
}

