% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeDistributionModels.R
\name{equalPop}
\alias{equalPop}
\title{Create equal colony sizes.}
\usage{
equalPop(nColonies, colonySize)
}
\arguments{
\item{nColonies}{The number of colonies. A single integer.}

\item{colonySize}{The size of all colonies}
}
\description{
Create equal colony sizes.
}

