% Generated by roxygen2 (4.0.1): do not edit by hand
\name{exponentialPop}
\alias{exponentialPop}
\title{Create exponential colony size distribution.}
\usage{
exponentialPop(nColonies, meanColonySize)
}
\arguments{
\item{nColonies}{The number of colonies. A single integer.}

\item{meanColonySize}{Mean colony size.}
}
\description{
Create random colony size distirbution from exponential distribution.
}
\seealso{
Other colonysize.distribution: \code{\link{equalPop}};
  \code{\link{poissonPop}}
}

