% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findCoinfDistr}
\alias{findCoinfDistr}
\title{Calculate coinfection disease distributions (how many individuals infected with number of diseases) at the end of a simulation}
\usage{
findCoinfDistr(pop, final = 10)
}
\arguments{
\item{pop}{A population object from makePop()}

\item{final}{Integer giving the number of recorded population states to average over. Note that this is the number of recorded events, not actual events so will cover a period of final * sample.}
}
\description{
Calculate coinfection disease distributions (how many individuals infected with number of diseases) at the end of a simulation
}

