% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{findDiseaseAdded}
\alias{findDiseaseAdded}
\title{Make vector of which disease is added for each coinfection transitions in pop$transitions}
\usage{
findDiseaseAdded(pop)
}
\arguments{
\item{pop}{A population object}
}
\description{
Make vector of which disease is added for each coinfection transitions in pop$transitions
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{deathR}};
  \code{\link{dispersalR}}; \code{\link{infectionR}},
  \code{\link{infectionR}}, \code{\link{recoveryR}};
  \code{\link{infectionTrans}};
  \code{\link{initTransitions}}; \code{\link{transRates}}
}

