% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findDiseaseAdded}
\alias{findDiseaseAdded}
\title{Make vector of which disease is added for each coinfection transitions in pop$transitions}
\usage{
findDiseaseAdded(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Make vector of which disease is added for each coinfection transitions in pop$transitions
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{deathR}};
  \code{\link{dispersalR}}; \code{\link{infectionR}};
  \code{\link{infectionTrans}}; \code{\link{transRates}}
}

