% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{infectionR}
\alias{infectionR}
\alias{recoveryR}
\title{Calculate infection rates for each colony.}
\usage{
infectionR(pop, t)

recoveryR(pop, t)
}
\arguments{
\item{pop}{A population object}

\item{t}{Time step. Note the population should be AT time t, going to t+1.}
}
\value{
nColonies * nPathogens vector Grouped by pathogen

nColonies * nPathogens vector Grouped by pathogen
}
\description{
Calculate infection rates for each colony.

Calculate recovery rates for each colony.
}

