% Generated by roxygen2 (4.0.1): do not edit by hand
\name{infectionR}
\alias{infectionR}
\title{Calculate infection rates for each colony.}
\usage{
infectionR(pop, t, infectTrans)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}

\item{t}{Time. An integer giving the event number (i.e. the 1000th event).}
}
\value{
nColonies * nPathogens vector Grouped by pathogen
}
\description{
Calculate infection rates for each colony.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{deathR}};
  \code{\link{dispersalR}}; \code{\link{findDiseaseAdded}};
  \code{\link{infectionTrans}};
  \code{\link{initTransitions}}; \code{\link{transRates}}
}

