% Generated by roxygen2 (4.0.1): do not edit by hand
\name{infectionTrans}
\alias{infectionTrans}
\title{Calculate infection rates for each colony.}
\usage{
infectionTrans(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Calculate starting infection rate for each colony (infection of susceptibles only).
  I guess I'm assuming that there's no multiinfection at this stage.
  However, as multiple pathogens can be seeded into same colony, not true.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{deathR}}; \code{\link{dispersalR}}
}

