% Generated by roxygen2 (4.0.2): do not edit by hand
\name{infectionTrans}
\alias{infectionTrans}
\title{Find possible possible infection transitions.}
\usage{
infectionTrans(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Find possible possible infection transitions.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{deathR}}; \code{\link{dispersalR}}
}

