% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{infectionTrans}
\alias{infectionTrans}
\title{Find possible possible infection transitions.}
\usage{
infectionTrans(pop)
}
\arguments{
\item{pop}{A population object}
}
\description{
Find possible possible infection transitions.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{deathR}};
  \code{\link{dispersalR}}; \code{\link{findDiseaseAdded}};
  \code{\link{infectionR}}, \code{\link{infectionR}},
  \code{\link{recoveryR}}; \code{\link{initTransitions}};
  \code{\link{transRates}}
}

