% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initTransitions}
\alias{initTransitions}
\title{Build the transition data.frame which will have rates populated later.}
\usage{
initTransitions(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Build the transition data.frame which will have rates populated later.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{deathR}};
  \code{\link{dispersalR}}; \code{\link{findDiseaseAdded}};
  \code{\link{infectionR}}; \code{\link{infectionTrans}};
  \code{\link{transRates}}
}

