% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makePop.R
\name{initTransitions}
\alias{initTransitions}
\title{Build the transition data.frame which will have rates populated later.}
\usage{
initTransitions(pop)
}
\arguments{
\item{pop}{A population object}
}
\description{
Build the transition data.frame which will have rates populated later.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{deathR}};
  \code{\link{dispersalR}}; \code{\link{findDiseaseAdded}};
  \code{\link{infectionR}}, \code{\link{infectionR}},
  \code{\link{recoveryR}}; \code{\link{infectionTrans}};
  \code{\link{transRates}}
}

