% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{inverseKern}
\alias{inverseKern}
\title{Inverse kernel for calculating a weight matrix from a distance matrix}
\usage{
inverseKern(distMatrix, thresh)
}
\arguments{
\item{distMatrix}{A distance matrix of the euclidean distances between each pair of colonies}

\item{thresh}{A maximum distance above which colonies are not connected.}
}
\description{
1/distance with a threshold.
}

