% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{MetapopEpi}
\alias{MetapopEpi}
\alias{MetapopEpi-package}
\title{A package containing functions to run multi pathogen, metapopulation epidemiological simulations.}
\description{
A package containing functions to run multi pathogen, metapopulation epidemiological simulations.
}
\author{
Tim CD Lucas
}

