% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\docType{package}
\name{MetapopEpi}
\alias{MetapopEpi}
\alias{MetapopEpi-package}
\title{A package containing functions to run multi pathogen, metapopulation epidemiological simulations.}
\description{
A package containing functions to run multi pathogen, metapopulation epidemiological simulations.
}
\author{
Tim CD Lucas
}

