% Generated by roxygen2 (4.0.1): do not edit by hand
\name{netDegree}
\alias{netDegree}
\title{Find the network degree of colonies}
\usage{
netDegree(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Find the network degree of colonies
}
\seealso{
Other Networks: \code{\link{adjMatrix}};
  \code{\link{checkIfGiant}}; \code{\link{edgeList}};
  \code{\link{netStrength}}; \code{\link{numComponents}};
  \code{\link{popMatrix}}; \code{\link{weightMatrix}}
}

