% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{netDegree}
\alias{netDegree}
\title{Find the network degree of colonies}
\usage{
netDegree(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Find the network degree of colonies
}

