% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{netStrength}
\alias{netStrength}
\title{Find the network strength of colonies}
\usage{
netStrength(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
The strength of a node is the sum of it's weighted edges.
}

