% Generated by roxygen2 (4.0.1): do not edit by hand
\name{netStrength}
\alias{netStrength}
\title{Find the network strength of colonies}
\usage{
netStrength(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
The strength of a node is the sum of it's weighted edges.
}
\seealso{
Other Networks: \code{\link{adjMatrix}};
  \code{\link{checkIfGiant}}; \code{\link{edgeList}};
  \code{\link{netDegree}}; \code{\link{numComponents}};
  \code{\link{popMatrix}}; \code{\link{weightMatrix}}
}

