% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pAll}
\alias{pAll}
\title{Plot something...}
\usage{
pAll(pop, start = 1, end = NULL, o = FALSE)
}
\arguments{
\item{pop}{A MetapopEpi class object}

\item{start}{What event to start plotting from.}

\item{end}{What event to end plotting from. If NULL, plot until end.}

\item{o}{Logical. Truncate y-axis.}
}
\description{
Plot something...
}

