% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pDis}
\alias{pDis}
\alias{pSI}
\title{Plot the total number of infected individuals with each disease}
\usage{
pDis(pop, o = FALSE)

pSI(pop, o = FALSE)
}
\arguments{
\item{pop}{A population object}

\item{o}{Logical whether to plot from the origin or not}
}
\description{
Plot the total number of infected individuals with each disease

Plot the total number of infected individuals and total Susceptible.
}

