% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pDis}
\alias{pDis}
\title{Plot the total number of infected individuals with each disease}
\usage{
pDis(pop, start = 1, end = NULL, o = FALSE)
}
\arguments{
\item{pop}{A MetapopEpi class object}

\item{start}{What event to start plotting from.}

\item{end}{What event to end plotting from. If NULL, plot until end.}

\item{o}{Logical. Truncate y-axis.}
}
\description{
Plot the total number of infected individuals with each disease
}

