% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pInf}
\alias{pInf}
\title{Plot the total number of infected individuals in each colony}
\usage{
pInf(pop, o = FALSE)
}
\arguments{
\item{pop}{A MetapopEpi class object}

\item{o}{Logical. Truncate y-axis.}
}
\description{
Plot the total number of infected individuals in each colony
}

