% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pInf}
\alias{pInf}
\alias{pInfold}
\title{Plot the total number of infected individuals in each colony}
\usage{
pInfold(pop, o = FALSE)

pInf(pop, start = 1, end = NULL, o = FALSE)
}
\arguments{
\item{pop}{A population object}

\item{o}{Logical whether to plot from the origin or not}

\item{start}{What event to start plotting from.}

\item{end}{What event to end plotting from. If NULL, plot until end.}
}
\description{
Plot the total number of infected individuals in each colony

Plot the total number of infected individuals in each colony
}

