% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotColonyNet}
\alias{colonyHeat}
\alias{plotColonyNet}
\title{Plot the population network in space}
\usage{
plotColonyNet(pop, lwd = 1.2, axes = FALSE, col = 1, area = NULL,
  alpha = 0.1, ...)

colonyHeat(weightMatrix)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}

\item{lwd}{Relative line width. This value is scaled to a reasonable value.}

\item{axes}{If set to FALSE, don't plot any axes}

\item{col}{Point colour}

\item{area}{The length of the sides of the plotting area. Either a length 2 numeric giving the x and y lengths, or a length one numeric giving the length of both.}

\item{alpha}{Transparency level in range (0, 1).}

\item{...}{Further arguments to \code{par}.}

\item{weightMatrix}{A matrix}
}
\description{
Plot showing the colonies, and the edges between them, with line thickness indicating weight.

Plot showing the weight of edges by colouring the adjacency matrix.
}

