% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotColonyNet}
\alias{colonyHeat}
\alias{plotColonyNet}
\title{Plot the population network in space}
\usage{
plotColonyNet(pop, lwd = 2)

colonyHeat(weightMatrix)
}
\arguments{
\item{lwd}{Relative line width. This value is scaled to a reasonable value.}

\item{weightMatrix}{A matrix}

\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Plot showing the colonies, and the edges between them, with line thickness indicating weight.

Plot showing the weight of edges by colouring the adjacency matrix.
}
\seealso{
Other viz: \code{\link{popHeat}}

Other viz: \code{\link{popHeat}}
}

