% Generated by roxygen2 (4.0.1): do not edit by hand
\name{poissonPop}
\alias{poissonPop}
\title{Create colony sizes based on poisson distribution.}
\usage{
poissonPop(nColonies, meanColonySize)
}
\arguments{
\item{nColonies}{The number of colonies. A single integer.}

\item{meanColonySize}{Mean colony size.}
}
\description{
Create colony sizes based on poisson distribution.
}
\seealso{
Other colonysize.distribution: \code{\link{equalPop}};
  \code{\link{exponentialPop}}
}

