% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeDistributionModels.R
\name{poissonPop}
\alias{poissonPop}
\title{Create colony sizes based on poisson distribution.}
\usage{
poissonPop(nColonies, meanColonySize)
}
\arguments{
\item{nColonies}{The number of colonies. A single integer.}

\item{meanColonySize}{Mean colony size.}
}
\description{
Create colony sizes based on poisson distribution.
}

