% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{popHeat}
\alias{popHeat}
\title{A nice heat map for weighted adjacency matrices}
\usage{
popHeat(pop)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}
}
\description{
Plot showing the weight of edges by colouring the adjacency matrix.
}

