% Generated by roxygen2 (4.0.1): do not edit by hand
\name{popMatrix}
\alias{popMatrix}
\title{Calculate a dispersal matrix}
\usage{
popMatrix(locations, maxDistance, kern)
}
\arguments{
\item{locations}{x, y numeric locations of each colony.}

\item{maxDistance}{For kernels with an upper threshold}

\item{kern}{Character, giving the name of the distance kernel to be used.}
}
\description{
Calulate a weighted (or unweighted) adjacency matrix. Then weight this value to convert to a dispersal matrix.
  An element i, j is the proportion of individuals dispersing from i that go to j.
}
\seealso{
Other Networks: \code{\link{adjMatrix}};
  \code{\link{checkIfGiant}}; \code{\link{edgeList}};
  \code{\link{netDegree}}; \code{\link{netStrength}};
  \code{\link{numComponents}}; \code{\link{weightMatrix}}
}

