% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pPop}
\alias{pPop}
\title{Plot the total number of individuals in each colony}
\usage{
pPop(pop, o = FALSE)
}
\arguments{
\item{pop}{A MetapopEpi class object}

\item{o}{Logical. Truncate y-axis.}
}
\description{
Plot the total number of individuals in each colony
}

