% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pSI}
\alias{pSI}
\title{Plot the total number of infected individuals and total Susceptible.}
\usage{
pSI(pop, o = FALSE)
}
\arguments{
\item{pop}{A population object}

\item{o}{Logical whether to plot from the origin or not}
}
\description{
Plot the total number of infected individuals and total Susceptible.
}

