% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pSI}
\alias{pSI}
\alias{pSIold}
\title{Plot the total number of infected individuals and total Susceptible.}
\usage{
pSIold(pop, o = FALSE)

pSI(pop, start = 1, end = NULL)
}
\arguments{
\item{pop}{A population object}

\item{o}{Logical whether to plot from the origin or not}

\item{start}{What event to start plotting from.}

\item{end}{What event to end plotting from. If NULL, plot until end.}
}
\description{
Plot the total number of infected individuals and total Susceptible.

Plot the total number of infected individuals and total Susceptible.
}

