% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pSI}
\alias{pSI}
\title{Plot the total number of infected individuals and total Susceptible.}
\usage{
pSI(pop, start = 1, end = NULL)
}
\arguments{
\item{pop}{A MetapopEpi class object}

\item{start}{What event to start plotting from.}

\item{end}{What event to end plotting from. If NULL, plot until end.}
}
\description{
Plot the total number of infected individuals and total Susceptible.
}

