% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{pSus}
\alias{pSus}
\title{Plot the number of susceptible individuals in each colony}
\usage{
pSus(pop, start = 1, end = NULL, o = FALSE)
}
\arguments{
\item{pop}{A population object}

\item{o}{Logical whether to plot from the origin or not}
}
\description{
Plot the number of susceptible individuals in each colony
}

