% Generated by roxygen2 (4.0.2): do not edit by hand
\name{randEvent}
\alias{randEvent}
\title{Perform a random event}
\usage{
randEvent(pop, t)
}
\arguments{
\item{pop}{A population object}

\item{t}{Time step. Note the population should be AT time t, going to t+1.}
}
\description{
Randomly selects an event weighted by their rates, runs the event and then calculates a waiting time.
}

