% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runSim}
\alias{runSim}
\title{Run simulation up to certain event number}
\usage{
runSim(pop, time = "end")
}
\arguments{
\item{t}{The event number to run the simulation until.}

\item{pop}{A population object}
}
\description{
Each time step, pick a random event and perform that event.
}

