% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{seedPathogen}
\alias{seedPathogen}
\title{Add single individuals infected with a number of pathogens}
\usage{
seedPathogen(pop, pathogens = 1, n = 1, diffCols = TRUE)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}

\item{pathogens}{Character vector of which pathogens should be seeded}

\item{n}{How many individuals should be infected.}

\item{diffCols}{Logical. If True, pathogen infections are forced to be in different colonies. Allows n to be as big as meanColonySize but requires nColonies > nPathogens}
}
\value{
An update object of same structure as from makePop
}
\description{
Allows addition of an infected individual of 1 or more pathogens.
}
\examples{
p <- makePop()
u <- seedPathogen(p, pathogens = 1:p$parameters['nPathogens'])
}

