% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sumI}
\alias{sumI}
\title{Returns an nColony x nPathogens matrix}
\usage{
sumI(pop, t)
}
\arguments{
\item{pop}{A metapopulation object as created by makePop().}

\item{t}{Time. An integer giving the event number (i.e. the 1000th event).}
}
\description{
To look at global infection, rather than colony-wise infection, this function sums across colonies
  to give an nColony x nPathogens for a certain time, t.
}

