% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transRates}
\alias{transRates}
\title{Calculate new transition rates.}
\usage{
transRates(pop, t)
}
\arguments{
\item{pop}{A population object}

\item{t}{Time step. Note the population should be AT time t, going to t+1.}
}
\description{
Calculate new transition rates.
}
\seealso{
Other initialRates: \code{\link{birthR}};
  \code{\link{coinfectionR}}; \code{\link{deathR}};
  \code{\link{dispersalR}}; \code{\link{findDiseaseAdded}};
  \code{\link{infectionR}}, \code{\link{infectionR}},
  \code{\link{recoveryR}}; \code{\link{infectionTrans}};
  \code{\link{initTransitions}}
}

