% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenDiversity.R
\name{transRates}
\alias{transRates}
\title{Calculate new transition rates.}
\usage{
transRates(pop, t)
}
\arguments{
\item{pop}{A MetapopEpi object}

\item{t}{Time. tMod + 1 so if time = 1000, sample = 1000, t here is 1}
}
\description{
Calculate new transition rates.
}

