% Generated by roxygen2 (4.0.1): do not edit by hand
\name{waitingTime}
\alias{waitingTime}
\title{Takes a number [0,1] (which is presumably random) and returns waiting time.}
\usage{
waitingTime(pop, t)
}
\arguments{
\item{randU}{Numeric in [0,1]. Typically this will a random uniform number.}
}
\description{
A single draw from a exponential distribution given randU, a number between 0 and 1.
  As population now has
}

